# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Posts",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "summary": "List of Posts",
                "operationId": "ed62344e51819929a660d340968cf4b5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "website_id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "title": "This is title for post #1",
                                        "content": "Post content #1",
                                        "author": "John Doe",
                                        "website_id": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new Post",
                "operationId": "1909fb5cb2c848593c1868c10bea4b73",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "website_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "This is title for post #1",
                                    "content": "Post content #1",
                                    "author": "John Doe",
                                    "website_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "id": 1,
                                            "title": "This is title for post #1",
                                            "content": "Post content #1",
                                            "author": "John Doe",
                                            "website_id": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Validation error.",
                                        "value": {
                                            "error": "Invalid input: website doesn't exists"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Server error.",
                                        "value": {
                                            "error": "Internal error: something went wrong, try again please"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "summary": "Get single post",
                "operationId": "76bfa01a5b6edb92a931e3ba48b8af5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "id": 1,
                                            "title": "This is title for post #1",
                                            "content": "Post content #1",
                                            "author": "John Doe",
                                            "website_id": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete single post",
                "operationId": "5933f33eb54123199703853572f150f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Server error.",
                                        "value": {
                                            "error": "Internal error: something went wrong, try again please"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/subscribe/{website_id}": {
            "post": {
                "summary": "Create new Website",
                "operationId": "c9fca27f12e4f26460eb62152e06fc71",
                "parameters": [
                    {
                        "name": "website_id",
                        "in": "path",
                        "description": "website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Website doesn't exists",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Validation error.",
                                        "value": {
                                            "error": "website exists"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}